requirements.txt
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.4
django-allauth==0.39.1
django-countries==5.4
django-crispy-forms==1.7.2
django-tinymce==2.8.0
django-tinymce4-lite==1.7.5
idna==2.8
jsmin==2.2.2
oauthlib==3.1.0
Pillow==6.1.0
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
sqlparse==0.3.0
stripe==2.35.1
urllib3==1.25.3
